Refactor Add to Metamask
