+++ src/cow-react/modules/wallet/api/pure/AddToMetamask/index.tsx
@@ -1,20 +1,12 @@
-import React, { useCallback, useContext, useState } from 'react'
+import React, { useContext } from 'react'
 import { Currency } from '@uniswap/sdk-core'
 import styled, { ThemeContext } from 'styled-components/macro'
 
-import { useWeb3React } from '@web3-react/core'
-import { getIsMetaMask } from '@cow/modules/wallet/api/utils'
 import { CheckCircle } from 'react-feather'
 import { RowFixed } from 'components/Row'
 import MetaMaskLogo from '../../assets/metamask.png'
-import { addTokenToMetamaskAnalytics } from 'components/analytics'
-import useCurrencyLogoURIs from 'lib/hooks/useCurrencyLogoURIs'
 import { TokenSymbol } from '@cow/common/pure/TokenSymbol'
 
-export type AddToMetamaskProps = {
-  currency: Currency | undefined
-  shortLabel?: boolean
-}
 
 export const ButtonCustom = styled.button`
   display: flex;
@@ -62,39 +54,16 @@ const CheckCircleCustom = styled(CheckCircle)`
   margin: 0 10px 0 0;
 `
 
-export default function AddToMetamask(props: AddToMetamaskProps) {
-  const { currency, shortLabel } = props
+export type AddToMetamaskProps = {
+  currency: Currency | undefined
+  shortLabel?: boolean
+  addToken: () => void
+  success?: boolean
+}
+
+export function AddToMetamask(props: AddToMetamaskProps) {
+  const { currency, shortLabel, addToken, success } = props
   const theme = useContext(ThemeContext)
-  const { connector } = useWeb3React()
-  const isMetamask = getIsMetaMask()
-
-  const [success, setSuccess] = useState<boolean | undefined>()
-
-  const token = currency?.wrapped
-  const logoURL = useCurrencyLogoURIs(token)[0]
-
-  const addToken = useCallback(() => {
-    if (!token?.symbol || !connector.watchAsset) return
-    connector
-      .watchAsset({
-        address: token.address,
-        symbol: token.symbol,
-        decimals: token.decimals,
-        image: logoURL,
-      })
-      .then(() => {
-        addTokenToMetamaskAnalytics('Succeeded', token.symbol)
-        setSuccess(true)
-      })
-      .catch(() => {
-        addTokenToMetamaskAnalytics('Failed', token.symbol)
-        setSuccess(false)
-      })
-  }, [connector, logoURL, token])
-
-  if (!currency || !isMetamask) {
-    return null
-  }
 
   return (
     <ButtonCustom onClick={addToken}>
