+++ src/cow-react/pages/Account/Balances.tsx
@@ -26,7 +26,7 @@ import { COW, V_COW } from 'constants/tokens'
 import { useErrorModal } from 'hooks/useErrorMessageAndModal'
 import { OperationType } from 'components/TransactionConfirmationModal'
 import useTransactionConfirmationModal from 'hooks/useTransactionConfirmationModal'
-import AddToMetamask from '@cow/modules/wallet/api/components/AddToMetamask'
+import AddToMetamask from '@cow/modules/wallet/web3-react/containers/AddToMetamask'
 import { Link } from 'react-router-dom'
 import CopyHelper from 'components/Copy'
 import { SwapVCowStatus } from 'state/cowToken/actions'
