+++ src/cow-react/pages/Account/styled.tsx
@@ -5,7 +5,7 @@ import { BannerExplainer } from '@cow/pages/Claim/styled'
 import * as CSS from 'csstype'
 import { transparentize } from 'polished'
 import { ExternalLink } from 'theme'
-import { ButtonCustom as AddToMetaMask } from '@cow/modules/wallet/api/components/AddToMetamask'
+import { ButtonCustom as AddToMetaMask } from '@cow/modules/wallet/web3-react/containers/AddToMetamask'
 import { CopyIcon as ClickToCopy } from 'components/Copy'
 import SVG from 'react-inlinesvg'
 import SpinnerLoader from 'components/Loader'
