+++ src/cow-react/pages/Claim/ClaimingStatus.tsx
@@ -18,7 +18,7 @@ import { Link } from 'react-router-dom'
 import { ExplorerLink } from 'components/ExplorerLink'
 import { EnhancedTransactionLink } from 'components/EnhancedTransactionLink'
 import { V_COW } from 'constants/tokens'
-import AddToMetamask from '@cow/modules/wallet/api/components/AddToMetamask'
+import AddToMetamask from '@cow/modules/wallet/web3-react/containers/AddToMetamask'
 import SVG from 'react-inlinesvg'
 import twitterImage from 'assets/cow-swap/twitter.svg'
 import discordImage from 'assets/cow-swap/discord.svg'
